User-Settings: move streamVideo functions to user-settings
